app/requirements.txt
flask==3.0.0
flask-cors==4.0.0
streamlit==1.29.0
pyodbc==5.0.1
pandas==2.1.4
plotly==5.18.0
python-dotenv==1.0.0
requests==2.31.0
